requirements.txt
Flask==2.3.2
flask-cors==3.0.10
requests==2.31.0
gunicorn==21.2.0
Pillow==10.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-SQLAlchemy 
psycopg2-binary
